requirements.txt
asgiref==3.8.1
Django==5.1
django-environ==0.11.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
Pillow>=8.0.0
django-ckeditor-5==0.2.17
django-import-export==2.7.0